refactor: tutorial model logic
